% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates_of_last_update.R
\name{get_dates_of_last_update}
\alias{get_dates_of_last_update}
\title{Get dates of last update of download files}
\usage{
get_dates_of_last_update(files = download_files())
}
\arguments{
\item{files}{file names to whose update date is to be checked in the
downloads page.}
}
\value{
Returns a tibble with one file per row.
}
\description{
This function will scrape on-the-fly GEISHA's downloads page and retrieve the
date of last update of the to-be-downloaded files.
}
