% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Get the files available to download from GEISHA}
\usage{
get_files(files = download_files())
}
\arguments{
\item{files}{The name of the files to be downloaded. This parameter is set by
default to download all available files.}
}
\description{
This function retrieves those files available for download from the
downloads' page of the GEISHA website. These files are saved to
\code{data-raw/} are meant to be used only by the package developer.
}
