% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_stages.R
\name{hh_stages}
\alias{hh_stages}
\title{Hamilton-Hamburguer stages}
\usage{
hh_stages(hh_prefix = FALSE)
}
\arguments{
\item{hh_prefix}{A logical indicating whether to include the prefix
\code{'HH'} in the after-laying egg stages, i.e., those stages numbered
with arabic numerals.}
}
\value{
A character vector of Hamilton-Hamburguer (HH) stages.
}
\description{
This function returns a character vector of Hamilton-Hamburguer (HH) stages.
}
