% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_updates}
\alias{last_updates}
\title{Date of last update of download files}
\format{
A \code{\link[tibble]{tibble}} of 2 variables:
\describe{
\item{\code{file}}{File name.}
\item{\code{date}}{Date of last update by GEISHA.}
}
}
\usage{
last_updates
}
\description{
Date of last update of download files
}
\keyword{datasets}
