% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_expression_txt.R
\name{read_expression_txt}
\alias{read_expression_txt}
\title{Imports data-raw/expression.txt as a tibble}
\usage{
read_expression_txt()
}
\value{
Returns a tibble with the data present in \code{expression.txt}.
}
\description{
This function reads \code{data-raw/expression.txt} as a tibble. This function
is meant to be used only by the developer of this package.
}
