% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_expression_xml.R
\name{read_expression_xml}
\alias{read_expression_xml}
\title{Imports data-raw/expression.xml as a tibble}
\usage{
read_expression_xml()
}
\value{
Returns a tibble with (part of) the data present in \code{expression.xml}.
}
\description{
This function reads \code{data-raw/expression.xml} as a tibble. This function
is meant to be used only by the developer of this package.
}
